requirements.txt
fastapi
urllib3
uvicorn
openai
qdrant-client
botocore
boto3 
python-dotenv
fastembed
numpy
pymongo
typing
datetime
bcrypt
PyJWT

llama-index-vector-stores-qdrant 
llama-index-readers-file 
llama-index-embeddings-fastembed 
llama-index-llms-openai
ipython